+++ src/Entity/Tenant/Theme.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\Entity\Tenant;
 
 use App\Entity\Traits\EntityTitleDescriptionTrait;
-use App\Entity\Traits\RelationsModifiedAtTrait;
 use App\Repository\ThemeRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
@@ -13,12 +12,9 @@ use Doctrine\ORM\Mapping as ORM;
 
 #[ORM\Entity(repositoryClass: ThemeRepository::class)]
 #[ORM\EntityListeners([\App\EventListener\ThemeDoctrineEventListener::class])]
-#[ORM\Index(fields: ['relationsModifiedAt'], name: 'relations_modified_at_idx')]
-#[ORM\Index(fields: ['modifiedAt'], name: 'modified_at_idx')]
 class Theme extends AbstractTenantScopedEntity
 {
     use EntityTitleDescriptionTrait;
-    use RelationsModifiedAtTrait;
 
     #[ORM\Column(type: \Doctrine\DBAL\Types\Types::TEXT)]
     private string $cssStyles = '';
@@ -87,7 +83,7 @@ class Theme extends AbstractTenantScopedEntity
         return $this->logo;
     }
 
-    public function addLogo(Media $medium): self
+    public function setLogo(?Media $medium): self
     {
         $this->logo = $medium;